moved celebrity detection related files to /backend/app/models/celebrity_detection folder
