+++ backend/app/models/celebrity_detection/CelebrityMatcher.py
